Add files via upload
I think I got everything. I edited the rng schema to make it such that the citation location wasn't included in the quote marks. It was easier than mucking about with following-sibling etc. to exclude the <cite> elements.
I had to guess on some honorifics. I think that any clergyperson is Revd. ? But I'm not quite sure.